Don't wrap the overridden element with a container
The wrapper causes problems with inputs that use :focus to, e.g., change
the underline border color or animate moving placeholder text via CSS
only. Without having the input as a sibling this can only be done with
javascript (unless you're only dealing with one input on the whole
page).